+++ changelog.md
@@ -1,5 +1,4 @@
 CHECK PREVIOUS PATCH NOTES!!
 
-- hotfix: fixed a crash due to compat mixins into other mods
-- fixed confetti popper not having cooldown
-- fixed sliding block not updating blocks around
+- fixed some tags
+- fixed a crash on neoforge